Tests/LibWeb/Text/expected/css-has-invalidation/cold-shadow-root-sheet-insights.txt

[cold shadow roots without :has()]
  hasInvalidationRuleCacheBuilds: 0
  hasAncestorWalkInvocations: 0
rgb(255, 0, 0)
[cold shadow root with :has()]
  hasInvalidationRuleCacheBuilds: 1
  hasAncestorWalkInvocations: 1
